services/stt/pyproject.toml
[project]
name = "ai-grpc-service"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "faster-whisper==1.2.1",
    "grpcio==1.80.0",
    "grpcio-tools==1.80.0",
    "psutil==7.2.2",
    "requests==2.33.1",
    "python-json-logger==4.1.0",
    "pyannote-audio==4.0.4",
    "ctranslate2==4.7.1",
    "nvidia-cuda-runtime-cu12; sys_platform == 'linux'",
    "nvidia-cublas-cu12; sys_platform == 'linux'",
    "nvidia-cudnn-cu12; sys_platform == 'linux'",
    "nvidia-curand-cu12; sys_platform == 'linux'",
    "nvidia-cuda-nvrtc-cu12; sys_platform == 'linux'",
    "nvidia-npp-cu12; sys_platform == 'linux'",
    "vllm>=0.19.0; sys_platform == 'linux'",
    "mlx-lm>=0.31.2; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchcodec>=0.10.0",
    "voyageai>=0.3.7",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest==9.0.3",
    "pytest-mock==3.15.1",
    "ruff==0.15.10",
    "types-grpcio>=1.0.0.20260408",
    "types-protobuf>=7.34.1.20260408",
    "types-psutil>=7.2.2.20260130",
    "types-requests>=2.33.0.20260408",
]

[tool.uv]
default-groups = ["dev"]
environments = [
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]
torchcodec = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."] # Ensures pytest resolves src.* imports correctly from the root

[tool.mypy]
files = ["src", "tests", "scripts"]
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "faster_whisper.*",
    "ctranslate2.*",
    "pyannote.*",
    "voyageai.*",
    "vllm.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["src.generated.*"]
ignore_errors = true